Merge main to branch

+++ src/controllers/join-event/delete.js
@@ -1,13 +1,15 @@
+const { isAuthorized } = require("../../utils/auth-utils");
+
 const deleteJoinEvent = async (req, res) => {
   try {
     const {
       session,
       db: { JoinEvent },
-      body: { eventId },
+      params: { userId, eventId },
     } = req;
-    console.log(session.userId, eventId);
-    const join = await JoinEvent.deleteEvent(session.userId, eventId);
-    res.status(201).json(join);
+    if (!isAuthorized(userId, session)) return res.sendStatus(403);
+    const del = await JoinEvent.deleteEvent(session.userId, eventId);
+    res.status(201).json(del);
   } catch (error) {
     console.error(error);
     res.status(500).json({ error: error.message });
